+++ TablePro/Models/TableFilter.swift
@@ -155,11 +155,15 @@ struct TabFilterState: Equatable, Codable {
     var filters: [TableFilter]
     var appliedFilters: [TableFilter]
     var isVisible: Bool
+    var quickSearchText: String
+    var filterLogicMode: FilterLogicMode
 
     init() {
         self.filters = []
         self.appliedFilters = []
         self.isVisible = false
+        self.quickSearchText = ""
+        self.filterLogicMode = .and
     }
 
     var hasChanges: Bool {